+++ packages/components/card/src/index.ts
@@ -1,4 +1,4 @@
 export { Card, type CardProps } from "./card";
-export { CardContent } from "./card-content";
-export { CardFooter } from "./card-footer";
+export { CardContent, type CardContentProps } from "./card-content";
+export { CardFooter, type CardFooterProps } from "./card-footer";
 export { CardHeader, type CardHeaderProps } from "./card-header";