+++ packages/components/dialog/src/index.ts
@@ -1,5 +1,5 @@
 export { Dialog, type DialogProps } from "./dialog";
-export { DialogBody } from "./dialog-body";
-export { DialogContent } from "./dialog-content";
-export { DialogFooter } from "./dialog-footer";
-export { DialogHeader } from "./dialog-header";
+export { DialogBody, type DialogBodyProps } from "./dialog-body";
+export { DialogContent, type DialogContentProps } from "./dialog-content";
+export { DialogFooter, type DialogFooterProps } from "./dialog-footer";
+export { DialogHeader, type DialogHeaderProps } from "./dialog-header";