+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.25
+
+### Patch Changes
+
+- Updated dependencies
+  - @jamsr-ui/react@0.0.25
+
 ## 0.1.24
 
 ### Patch Changes